DISTINGUISH create_disp_day (Python) from create_disp_dat_fast (C)
