+++ .github/workflows/c-cpp.yml
@@ -14,11 +14,11 @@ jobs:
         working-directory: ./fem/post
     steps:
     - uses: actions/checkout@v2
-    - name: generate create_disp_dat and scale_disp_dat Makefile with cmake
+    - name: generate create_disp_dat_fast and scale_disp_dat Makefile with cmake
       run: cmake .
     - name: make targets
-      run: make
-    - name: generate create_disp_dat disp.dat output
-      run: ./create_disp_dat -i ../../tests/nodout1 -o ./disp.dat
+      run: cmake --build .
+    - name: generate create_disp_dat_fast disp.dat output
+      run: ./create_disp_dat_fast -i ../../tests/nodout1 -o ./disp.dat
     - name: compare test output
       run: cmp ../../tests/disp.dat.good ./disp.dat
