+++ .github/workflows/python-package.yml
@@ -21,11 +21,6 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
-    - name: build modules with cmake
-      working-directory: ./fem/post
-      run: |
-        cmake .
-        make
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
