--- build_sphinx.sh
@@ -1,1 +1,0 @@
-PYTHONPATH=.. sphinx-build -nT -b dummy docs/ docs/_build/html
