--- fem/CMakeLists.txt
@@ -1,3 +1,0 @@
-cmake_minimum_required(VERSION 3.2)
-project(fem)
-add_subdirectory(post)
