+++ fem/post/.gitignore
@@ -1,3 +1,3 @@
+create_disp_dat_fast
+create_disp_dat_fast.py
 Makefile
-create_disp_dat
-create_disp_dat.py
