+++ fem/post/CMakeLists.txt
@@ -3,7 +3,7 @@ project(fem)
 
 set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS}")
 
-set(SOURCE_FILES create_disp_dat.c)
+set(SOURCE_FILES create_disp_dat_fast.c)
 
 FIND_PACKAGE(SWIG REQUIRED)
 INCLUDE(${SWIG_USE_FILE})
@@ -15,11 +15,11 @@ INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR})
 
 SET(CMAKE_SWIG_FLAGS "")
 
-SET_SOURCE_FILES_PROPERTIES(create_disp_dat.i PROPERTIES SWIG_FLAGS "-includeall" SWIG_MODULE_NAME create_disp_dat.so)
-SWIG_ADD_LIBRARY(create_disp_dat LANGUAGE python TYPE MODULE SOURCES create_disp_dat.i create_disp_dat.c)
-SWIG_LINK_LIBRARIES(create_disp_dat ${PYTHON_LIBRARIES})
+SET_SOURCE_FILES_PROPERTIES(create_disp_dat_fast.i PROPERTIES SWIG_FLAGS "-includeall" SWIG_MODULE_NAME create_disp_dat_fast.so)
+SWIG_ADD_LIBRARY(create_disp_dat_fast LANGUAGE python TYPE MODULE SOURCES create_disp_dat_fast.i create_disp_dat_fast.c)
+SWIG_LINK_LIBRARIES(create_disp_dat_fast ${PYTHON_LIBRARIES})
 
-add_executable(create_disp_dat ${SOURCE_FILES})
+add_executable(create_disp_dat_fast ${SOURCE_FILES})
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
 set(SCALE_DISP_DAT_SOURCE_FILES scale_disp_dat.cpp)
